+++ Modules/Application/Controller/d3GtmAccountWishlistController.php
@@ -4,7 +4,7 @@ namespace D3\GoogleAnalytics4\Modules\Application\Controller;
 
 class d3GtmAccountWishlistController extends d3GtmAccountWishlistController_parent
 {
-    protected $_sThisTemplate = 'page/account/d3gtmwishlist.tpl';
+    protected $_sThisTemplate = '@' . Constants::OXID_MODULE_ID . '/page/account/d3gtmwishlist.tpl';
 
     public function render()
     {